[ADD] ERP API 추가

+++ src/main/java/livart/erp/domain/promotion/dto/request/CouponRegisterRequest.java
@@ -4,6 +4,7 @@ import jakarta.persistence.Lob;
 import livart.common.dto.enums.coupon.*;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -20,7 +21,7 @@ public class CouponRegisterRequest {
     private LocalDate expireEndDate;
     private Long issuedDate;
     private CouponDiscountType couponDiscountType;
-    private Long discountPrice;
+    private BigDecimal discountPrice;
     private Boolean expiredMessage;
     private Integer minOrderPrice;
 }
